% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burstBounds.R
\name{burstBounds}
\alias{burstBounds}
\title{Find or define boundaries of sampling bursts in time series}
\usage{
burstBounds(x, Fs, burstLength = NULL)
}
\arguments{
\item{x}{A vector of POSIXct time stamps}

\item{Fs}{Sampling rate, units of Hertz (samples per second)}

\item{burstLength}{Desired length, in minutes, of sampling bursts to be analyzed}
}
\value{
A vector of row indices from input vector x that define time
chunks to be analyzed. Each value is the start and end of a time chunk
(except the 1st and last values of the vector)
}
\description{
Find or define boundaries of sampling bursts in a time series
}
\details{
This function searches for natural breaks in a time series, or
defines a set of breaks based on the desired length of sampling
bursts to be analyzed.
}
