% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinOWHLfiles.R
\name{joinOWHLfiles}
\alias{joinOWHLfiles}
\title{Concatenate multiple OWHL raw csv data files}
\usage{
joinOWHLfiles(filenames, timezone = "UTC", verbose = FALSE)
}
\arguments{
\item{filenames}{A vector of OWHL csv filenames, including path.}

\item{timezone}{Specifies the time zone the raw data timestamps represent. UTC is
the preferred time zone for simplicity.}

\item{verbose}{Logical argument TRUE or FALSE, specifying if progress messages should be output.}
}
\value{
A data frame with the individual csv files concatenated together
}
\description{
Concatenate multiple OWHL raw csv data files
}
\details{
Concatenate multiple OWHL raw csv data files into a single data frame and
interpolate any missing 1-second intervals. The resulting data frame can
be used to generate wave statistics from other functions in the package.

The input csv files can have mission info in the first row, and then the
following columns: POSIXt, DateTime, frac.seconds, Pressure.mbar, TempC
}
