% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBursts.R
\name{processBursts}
\alias{processBursts}
\title{Generate wave statistics for multiple sampling bursts}
\usage{
processBursts(Ht, times, burstLength = NULL, Fs, method = "sp",
  tzone = NULL)
}
\arguments{
\item{Ht}{A vector of sea surface heights}

\item{times}{A vector of POSIXct time stamps}

\item{burstLength}{The desired sampling burst length, in minutes}

\item{Fs}{The sampling rate of the data, units of Hz (samples per second)}

\item{method}{The wave statistic estimation method, either 'sp' for spectral
(default) or 'zc' for zero-crossing}

\item{tzone}{The time zone for input POSIXct time values. Default = 'UTC'}
}
\value{
A data frame of wave statistics (height and period) for each sampling
burst. The time value represents the time stamp at the end of the sampling
burst
}
\description{
Generate wave statistics for multiple sampling bursts
}
